requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.1.0
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
groq>=0.4.0
chromadb>=0.4.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
protobuf<=3.20.3
pysqlite3-binary
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5

# Optional (uncomment if needed)
# langchain-openai>=0.0.5
# langchain-cohere>=0.0.5
# openai>=1.10.0
# pdfplumber>=0.10.0
